# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Api Rest Adoptame",
    "description": "Documentación_",
    "version": "1.0.0"
  },
  "host": "localhost:5000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/api/v1.0/pets/": {
      "get": {
        "tags": [
          "PETS"
        ],
        "description": "",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "PETS"
        ],
        "description": "",
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "photos",
            "in": "formData",
            "type": "file",
            "required": false,
            "description": "Selecciona una foto",
            "collectionFormat": "multi",
            "items": {
              "type": "file"
            }
          },
          {
            "name": "body",
            "in": "body",
            "description": "Some description...",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "user_test"
                },
                "typeId": {
                  "type": "string",
                  "example": "dog"
                },
                "breedId": {
                  "type": "number",
                  "example": 2
                },
                "coat": {
                  "type": "string",
                  "example": "short"
                },
                "specialCares": {
                  "type": "boolean",
                  "example": false
                },
                "castrated": {
                  "type": "boolean",
                  "example": false
                },
                "gender": {
                  "type": "string",
                  "example": "male"
                },
                "environment": {
                  "type": "object",
                  "properties": {
                    "children": {
                      "type": "boolean",
                      "example": true
                    },
                    "dogs": {},
                    "cats": {}
                  }
                },
                "tags": {
                  "type": "array",
                  "example": [
                    "friendly",
                    "affectionate"
                  ],
                  "items": {
                    "type": "string"
                  }
                },
                "size": {
                  "type": "string",
                  "example": "medium"
                },
                "color": {
                  "type": "string",
                  "example": "marron"
                },
                "age": {
                  "type": "string",
                  "example": "young"
                },
                "health": {
                  "type": "string",
                  "example": "vaccinations up to date"
                },
                "description": {
                  "type": "string",
                  "example": "happy dog"
                },
                "userId": {
                  "type": "number",
                  "example": 1
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/v1.0/pets/{id}": {
      "get": {
        "tags": [
          "PETS"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "PETS"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "delete": {
        "tags": [
          "PETS"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  }
}